+++ lib/pplumbing_err/test/cram/main_climate.ml
@@ -4,4 +4,9 @@
 (*  SPDX-License-Identifier: MIT                                                 *)
 (*********************************************************************************)
 
-let () = Cmdlang_cmdliner_runner.run Test_command.main ~name:"main" ~version:"%%VERSION%%"
+let () =
+  Pplumbing_cmdlang_climate_runner.Cmdlang_climate_runner.run
+    Test_command.main
+    ~name:"main-climate"
+    ~version:"%%VERSION%%"
+;;